Update the README
Some more things need to be done:

How come the default doesn't have cardmodels as we do with notification? Can we make it consistent?
We need a new link for initialization customizations. I moved some of the more esoteric customizations out of the README to keep it short.
I removed Deployment; we should add that to the root README.md