fix(sdk-react): Clear needConsent flag after successful login (#6230)
When UIRequiredError occurs, needConsent is set to true so it calls teamsfx.login() at the next reload to ask for user consent. So far so good.
The problem is: after the user gives the necessary consent, needConsent flag is not cleared, so subsequent reloads will keep calling teamsfx.login and showing popup window that will automatically disappear.

This change fixes the issue - when consent is granted, login function won't be called.